+++ src/test/java/com/programmers/vouchermanagement/wallet/JdbcWalletRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.programmers.vouchermanagement.wallet;
 
 import com.programmers.vouchermanagement.VoucherManagementApplication;
-import com.programmers.vouchermanagement.global.common.JdbcRepositoryManager;
+import com.programmers.vouchermanagement.global.utils.UuidUtilsManager;
 import com.programmers.vouchermanagement.wallet.domain.Wallet;
 import com.programmers.vouchermanagement.wallet.repository.JdbcWalletRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -33,9 +33,9 @@ public class JdbcWalletRepositoryTest {
 
     private static final RowMapper<Wallet> walletRowMapper = (resultSet, index) -> {
 
-        UUID walletId = JdbcRepositoryManager.bytesToUUID(resultSet.getBytes("wallet_id"));
-        UUID customerId = JdbcRepositoryManager.bytesToUUID(resultSet.getBytes("customer_id"));
-        UUID voucherId = JdbcRepositoryManager.bytesToUUID(resultSet.getBytes("voucher_id"));
+        UUID walletId = UuidUtilsManager.bytesToUUID(resultSet.getBytes("wallet_id"));
+        UUID customerId = UuidUtilsManager.bytesToUUID(resultSet.getBytes("customer_id"));
+        UUID voucherId = UuidUtilsManager.bytesToUUID(resultSet.getBytes("voucher_id"));
 
         return new Wallet(walletId, customerId, voucherId);
     };
